requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
pymongo
python-dotenv
bcrypt
PyJWT

